Move fragment management out from ImpalaServer
ImpalaServer is still pretty big. One reason was that it was conflating
several separate responsibilities, including management of query
execution, management of fragment execution and responding directly to
Thrift RPCs.

This patch moves fragment execution into a separate FragmentMgr class,
and adds a (now-needed) proxy class to forward ImpalaInternalService
RPCs to either ImpalaServer or the FragmentMgr. Fragment management is
pretty self-contained (look at how small the resulting class had to be),
and so it was a good candidate to be broken out of ImpalaServer.

Change-Id: I24efb1a71a4225f9f85610565bf557a2e677e629
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4854
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins